requirements.txt
fastcore
onnxruntime==1.8.1
pqdm
fire
pyctcdecode
pypi-kenlm